Fix build errors in Preloader and Reset Password page
This commit addresses two separate build errors:

1.  In `components/Preloader.tsx`, the `useEffect` hook was used without being imported, causing a compilation failure. This has been fixed by adding `useEffect` to the import statement from 'react'.

2.  The `/reset-password` page was failing to build during prerendering because `useSearchParams()` was used in a way that is not compatible with Next.js static generation. The page has been refactored to move the logic that uses this hook into a separate client component, which is then wrapped in a `<Suspense>` boundary, as recommended by the Next.js documentation.
