+++ components/Preloader.tsx
@@ -1,7 +1,7 @@
 // components/Preloader.tsx
 "use client";
 
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import { motion, AnimatePresence } from 'framer-motion';
 import { useTranslation } from '@/context/LanguageContext';
 import { Button } from './ui/button';
